+++ src/__tests__/logdrains.test.ts
@@ -98,9 +98,7 @@ test("Logdrains Get Integration Log Drains", async () => {
       branch: "feature/*",
       samplingRate: 0.5,
       source: {
-        kind: "integration",
-        integrationId: "<id>",
-        integrationConfigurationId: "<id>",
+        kind: "self-served",
       },
     },
   ]);
@@ -153,7 +151,9 @@ test("Logdrains Create Log Drain", async () => {
     branch: "feature/*",
     samplingRate: 0.5,
     source: {
-      kind: "self-served",
+      kind: "integration",
+      integrationId: "<id>",
+      integrationConfigurationId: "<id>",
     },
   });
 });