+++ src/__tests__/mockserver/internal/handler/pathgetv10projects.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"mockserver/internal/handler/assert"
 	"mockserver/internal/logging"
+	"mockserver/internal/sdk/models/components"
 	"mockserver/internal/sdk/models/operations"
 	"mockserver/internal/sdk/types"
 	"mockserver/internal/sdk/utils"
@@ -52,9 +53,9 @@ func testGetProjectsGetProjects0(w http.ResponseWriter, req *http.Request) {
 				operations.GetProjectsProject2{
 					AccountID: "<id>",
 					Crons: &operations.GetProjectsCrons{
-						EnabledAt:    2103.82,
-						DisabledAt:   types.Float64(3154.28),
-						UpdatedAt:    3637.11,
+						EnabledAt:    8700.13,
+						DisabledAt:   types.Float64(7991.59),
+						UpdatedAt:    4614.79,
 						DeploymentID: types.String("<id>"),
 						Definitions:  []operations.GetProjectsDefinition{},
 					},
@@ -62,9 +63,13 @@ func testGetProjectsGetProjects0(w http.ResponseWriter, req *http.Request) {
 					DirectoryListing:     false,
 					ID:                   "<id>",
 					Name:                 "<value>",
-					NodeVersion:          operations.ProjectNodeVersion2TwentyFourDotX,
+					NodeVersion:          operations.ProjectNodeVersion2TwentyTwoDotX,
 					ResourceConfig: operations.ProjectResourceConfig2{
-						FunctionDefaultRegions: []string{},
+						FunctionDefaultRegions: []string{
+							"<value 1>",
+							"<value 2>",
+							"<value 3>",
+						},
 					},
 					RollingRelease: &operations.ProjectRollingRelease2{
 						Target: "production",
@@ -85,10 +90,11 @@ func testGetProjectsGetProjects0(w http.ResponseWriter, req *http.Request) {
 					},
 				},
 			},
-			Pagination: operations.CreatePaginationUnion2GetProjectsPagination2(
-				operations.GetProjectsPagination2{
+			Pagination: operations.CreatePaginationUnion2Pagination(
+				components.Pagination{
 					Count: 20,
-					Next:  types.String("JBSWY3DPEHPK3PXP"),
+					Next:  types.Float64(1540095775951),
+					Prev:  types.Float64(1540095775951),
 				},
 			),
 		},