+++ src/__tests__/mockserver/internal/handler/pathgetv6domainsdomainconfig.go
@@ -46,14 +46,20 @@ func testGetDomainConfigGetDomainConfig0(w http.ResponseWriter, req *http.Reques
 		return
 	}
 	var respBody *operations.GetDomainConfigResponseBody = &operations.GetDomainConfigResponseBody{
-		ConfiguredBy: operations.GetDomainConfigConfiguredByHTTP.ToPointer(),
-		AcceptedChallenges: []operations.AcceptedChallenge{
-			operations.AcceptedChallengeHttp01,
+		ConfiguredBy:       operations.GetDomainConfigConfiguredByDns01.ToPointer(),
+		AcceptedChallenges: []operations.AcceptedChallenge{},
+		RecommendedIPv4: []operations.RecommendedIPv4{
+			operations.RecommendedIPv4{
+				Rank: 5288.95,
+				Value: []string{
+					"<value 1>",
+					"<value 2>",
+				},
+			},
 		},
-		RecommendedIPv4: []operations.RecommendedIPv4{},
 		RecommendedCNAME: []operations.RecommendedCNAME{
 			operations.RecommendedCNAME{
-				Rank:  4375.87,
+				Rank:  710.36,
 				Value: "<value>",
 			},
 		},