+++ src/models/getteammembersop.ts
@@ -122,6 +122,7 @@ export const GetTeamMembersOrigin = {
   NsnbRequestAccess: "nsnb-request-access",
   NsnbViewerUpgrade: "nsnb-viewer-upgrade",
   NsnbInvite: "nsnb-invite",
+  NsnbRedeploy: "nsnb-redeploy",
 } as const;
 export type GetTeamMembersOrigin = ClosedEnum<typeof GetTeamMembersOrigin>;